# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "summary": "Register",
        "description": "Register user and return Bearer token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignupDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Guest token generaed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthTokenResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "409": {
            "description": "error: Email already in use",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        },
        "tags": ["auth"]
      }
    },
    "/auth/signin": {
      "post": {
        "operationId": "AuthController_signin",
        "summary": "Login",
        "description": "Sign in user and return Bearer token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SigninDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Guest token generated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthTokenResponse" }
              }
            }
          },
          "400": {
            "description": "Invalid Credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        },
        "tags": ["auth"]
      }
    },
    "/myAccount": {
      "get": {
        "operationId": "GuestController_getAccountInfo",
        "summary": "User info",
        "description": "Return guest account info with bearer token",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return guest info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestAccountResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        },
        "tags": ["myAccount"],
        "security": [{ "bearer": [] }]
      },
      "put": {
        "operationId": "GuestController_updateGuestInfo",
        "summary": "User info",
        "description": "Edit user account name or email",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Guest info updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UpdateUserDto" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        },
        "tags": ["myAccount"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "GuestController_DeleteGuestAccount",
        "summary": "User info",
        "description": "Delete user account using Bearer Token",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Delete guest account",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeletedResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        },
        "tags": ["myAccount"],
        "security": [{ "bearer": [] }]
      }
    },
    "/myAccount/reservations": {
      "get": {
        "operationId": "GuestController_getUserReservations",
        "summary": "User info",
        "description": "Return a list of user's bookings using Bearer token",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return guest bookings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestReservationsDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "No bookings found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        },
        "tags": ["myAccount"],
        "security": [{ "bearer": [] }]
      }
    },
    "/booking": {
      "get": {
        "operationId": "BookingController_getBookings",
        "summary": "Booking",
        "description": "Get all bookings in specified date range",
        "parameters": [
          {
            "name": "checkinDate",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "checkoutDate",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return Bookings in specified date",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BookingResponseDto" }
              }
            }
          },
          "400": {
            "description": "Incorrect date params",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "No bookings found in specified date",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        },
        "tags": ["booking"]
      },
      "post": {
        "operationId": "BookingController_createBooking",
        "summary": "Booking",
        "description": "Create a booking",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateBookingDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Return Room by id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBookingResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Sorry, this room is full",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        },
        "tags": ["booking"],
        "security": [{ "bearer": [] }]
      }
    },
    "/room": {
      "get": {
        "operationId": "RoomController_getRooms",
        "summary": "Room",
        "description": "Return all rooms available in the specified date range (yyyy-mm-dd). Date format example: \"2023-10-20\"",
        "parameters": [
          {
            "name": "checkinDate",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "checkoutDate",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all available Rooms in specified date",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RoomResponseDto" }
              }
            }
          },
          "400": {
            "description": "Incorrect date params",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "No rooms found in specified date",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        },
        "tags": ["rooms"]
      }
    },
    "/room/{id}": {
      "get": {
        "operationId": "RoomController_getRoomById",
        "summary": "Room Id",
        "description": "Return a room by id number",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return Room by id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RoomResponseDto" }
              }
            }
          },
          "404": {
            "description": "Room not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        },
        "tags": ["rooms"]
      }
    }
  },
  "info": {
    "title": "Athena Hotel",
    "description": "This is the API for Athena Hotel reservation system.",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string", "example": "email@example.com" },
          "password": { "type": "string" }
        },
        "required": ["name", "email", "password"]
      },
      "AuthTokenResponse": {
        "type": "object",
        "properties": { "token": { "type": "string" } },
        "required": ["token"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "error": { "type": "string" },
          "statusCode": { "type": "number" }
        },
        "required": ["message", "error", "statusCode"]
      },
      "SigninDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": ["email", "password"]
      },
      "GuestAccountResponseDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string", "example": "email@example.com" },
          "created_at": { "format": "date-time", "type": "string" }
        },
        "required": ["name", "email", "created_at"]
      },
      "GuestReservationsDto": {
        "type": "object",
        "properties": {
          "bookings": {
            "example": [{ "id": 1, "room_id": 1 }],
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["bookings"]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string", "example": "email@example.com" }
        },
        "required": ["name", "email"]
      },
      "DeletedResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Account deleted succesessfully"
          }
        },
        "required": ["message"]
      },
      "BookingResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "room_id": { "type": "number" },
          "date_check_in": { "format": "date-time", "type": "string" },
          "date_check_out": { "format": "date-time", "type": "string" }
        },
        "required": ["id", "room_id", "date_check_in", "date_check_out"]
      },
      "CreateBookingDto": {
        "type": "object",
        "properties": {
          "checkinDate": { "type": "string", "example": "2023-10-20" },
          "checkoutDate": { "type": "string", "example": "2023-10-25" },
          "guestNumber": { "type": "number", "example": 3 },
          "roomId": { "type": "number", "example": 2 }
        },
        "required": ["checkinDate", "checkoutDate", "guestNumber", "roomId"]
      },
      "CreateBookingResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "example": "2023-10-15"
          },
          "date_check_in": {
            "format": "date-time",
            "type": "string",
            "example": "2023-10-20"
          },
          "date_check_out": {
            "format": "date-time",
            "type": "string",
            "example": "2023-10-25"
          },
          "number_of_guests": { "type": "number", "example": 3 },
          "guest_id": { "type": "string" },
          "room_id": { "type": "number", "example": 1 }
        },
        "required": [
          "id",
          "created_at",
          "date_check_in",
          "date_check_out",
          "number_of_guests",
          "guest_id",
          "room_id"
        ]
      },
      "RoomResponseDto": {
        "type": "object",
        "properties": {
          "id": { "type": "number", "example": 1 },
          "quantity": { "type": "number", "example": 3 },
          "price": { "type": "number", "example": 4500 },
          "roomType": {
            "type": "string",
            "enum": [
              "SIMPLE_SINGLE",
              "PREMIUM_SINGLE",
              "SIMPLE_COUPLE",
              "PREMIUM_COUPLE"
            ]
          },
          "description": { "type": "string" },
          "size": { "type": "number", "example": 70 },
          "bed_size": { "type": "string", "example": "120x70" },
          "capacity": { "type": "number", "example": 2 },
          "floor": { "type": "number", "example": 12 }
        },
        "required": [
          "id",
          "quantity",
          "price",
          "roomType",
          "description",
          "size",
          "bed_size",
          "capacity",
          "floor"
        ]
      }
    }
  }
}